+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/mode/iconpack/common/model/PackEditState.kt
@@ -13,7 +13,6 @@ data class InputFieldState(
     val packageName: InputState,
     val nestedPacks: List<NestedPack>,
     val iconPackPackage: InputState = packageName,
-    val allowAddNestedPack: Boolean = true,
 ) {
     val isValid: Boolean = iconPackName.isValid() &&
         packageName.isValid() &&