+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/mode/iconpack/newpack/ui/model/NewPackAction.kt
@@ -12,6 +12,4 @@ sealed interface NewPackAction {
 
     data object PreviewPackObject : NewPackAction
     data object SavePack : NewPackAction
-
-    data class UseMaterialPack(val value: Boolean) : NewPackAction
 }